no app: desktop theme
Helper script that allows you to view the desktop theme in the app.  This script is only added to the desktop theme and removes a cookie AppPresser_Bypass that tell the plugin not to use the app theme. That cookie needs to be removed so the the app views other pages, those page do show the app theme.