cookie hack to allow login via browser preview
This is a fix because API based login does not set a cookie properly when using the browser preview. It creates an invisible iframe in the login modal that sets the cookie using a one-time-use auth key. This only happens in the browser, device login works fine without this hack.